Clever Claire working ok. Needs more tests in one part, improvement in another
3 hours

In a game, it's hard to reach a state where Claire runs out of
intersections to fire at, and that makes the logic for her splitting
plain gaps hard to test.

It might be better all around if she chose gap intersections to shoot
having maximum distance from any other shot space on the board.  I think
this would dice up the board a little faster

NOTE: 'hit' is a bad property name on cell.  It should really be 'shot'.
We don't call a miss in the ocean a hit when we verbally play this game.

NOTE 2: This commit also fixes a previous flaw in the chasing logic