Working strategy for Clever Claire (3 hours)
I'm tired, and I feel vaguely certain this algorithm is needlessly
complex.  I know it's not implemented in *efficient* terms, but that's
not my first objective.  I'm aiming for elegance first, and I'm not
sure I've achieved that.  Claire should never, ever fire at a spot in
a gap that's not big enough to hold one of the boats she's looking for.
I think after some light testing (need more) that I've achieved that much.